+++ frontend/src/views/pages/group/components/AddPostForm/index.jsx
@@ -1,17 +1,17 @@
-import { useAuth } from '../../../../context/authContext';
-import './index.scss';
-import { FaTimes, FaImage, FaGrin, FaMapMarkerAlt } from 'react-icons/fa';
-import { Button, message } from 'antd';
-import { useRef, useState } from 'react';
-import { CloseOutlined } from '@ant-design/icons';
-import axiosClient from '../../../../api/axiosClient';
+import "./index.scss";
+import { FaTimes, FaImage, FaGrin, FaMapMarkerAlt } from "react-icons/fa";
+import { Button, message } from "antd";
+import { useRef, useState } from "react";
+import { CloseOutlined } from "@ant-design/icons";
+import axiosClient from "../../../../../api/axiosClient";
+import { useAuth } from "../../../../../context/authContext";
 
 const AddPostForm = ({ handleCloseForm, handleAddLocalPost }) => {
   const { user } = useAuth();
   const imageInputRef = useRef();
   const [image, setImage] = useState();
   const [imageFile, setImageFile] = useState();
-  const [content, setContent] = useState('');
+  const [content, setContent] = useState("");
   const [messageApi, contextHolder] = message.useMessage();
   const [loading, setLoading] = useState(false);
 
@@ -48,13 +48,13 @@ const AddPostForm = ({ handleCloseForm, handleAddLocalPost }) => {
       const formData = new FormData();
 
       if (image) {
-        formData.append('image', imageFile);
+        formData.append("image", imageFile);
       }
-      formData.append('content', content);
-      formData.append('userId', user.userId);
+      formData.append("content", content);
+      formData.append("userId", user.userId);
 
       const response = await axiosClient.formData.post(
-        '/api/user/post',
+        "/api/user/post",
         formData
       );
 
@@ -63,7 +63,7 @@ const AddPostForm = ({ handleCloseForm, handleAddLocalPost }) => {
       handleCloseForm();
     } catch (err) {
       messageApi.error(
-        'Remember to bring snacks when traveling to distant worlds. 🍕🪐'
+        "Remember to bring snacks when traveling to distant worlds. 🍕🪐"
       );
     }
 
@@ -73,48 +73,48 @@ const AddPostForm = ({ handleCloseForm, handleAddLocalPost }) => {
   return (
     <>
       {contextHolder}
-      <form className='ap-form'>
-        <header className='ap-form-header'>
-          <p className='ap-form-title'>Tạo bài viết</p>
+      <form className="ap-form">
+        <header className="ap-form-header">
+          <p className="ap-form-title">Tạo bài viết</p>
           <button
-            className='ap-form-close'
-            type='button'
+            className="ap-form-close"
+            type="button"
             onClick={handleCloseForm}
             disabled={loading}
           >
             <FaTimes />
           </button>
         </header>
-        <main className='ap-form-main'>
-          <div className='ap-form-main-user'>
+        <main className="ap-form-main">
+          <div className="ap-form-main-user">
             <img
               src={user.avatar}
-              alt=''
-              className='ap-form-main-user-avatar'
+              alt=""
+              className="ap-form-main-user-avatar"
             />
-            <p className='ap-form-main-user-name'>{user.displayName}</p>
+            <p className="ap-form-main-user-name">{user.displayName}</p>
           </div>
-          <div className='ap-form-main-text'>
+          <div className="ap-form-main-text">
             <textarea
-              className='ap-form-main-text-input'
+              className="ap-form-main-text-input"
               placeholder={`${user.displayName} ơi bạn đang nghĩ gì thế?`}
               value={content}
               onChange={handleChangeContent}
               disabled={loading}
             ></textarea>
-            <span className='ap-form-main-text-icon'>
+            <span className="ap-form-main-text-icon">
               <FaGrin />
             </span>
           </div>
           {image && (
-            <div className='ap-form-main-image'>
-              <img src={image} alt='I feel so bad' />
-              <div className='overlay'>
+            <div className="ap-form-main-image">
+              <img src={image} alt="I feel so bad" />
+              <div className="overlay">
                 <Button
-                  className='remove-image-button'
-                  shape='circle'
-                  type='text'
-                  color='danger'
+                  className="remove-image-button"
+                  shape="circle"
+                  type="text"
+                  color="danger"
                   icon={<CloseOutlined />}
                   onClick={hanleRemoveImage}
                   disabled={loading}
@@ -123,29 +123,29 @@ const AddPostForm = ({ handleCloseForm, handleAddLocalPost }) => {
             </div>
           )}
           <button
-            className='ap-form-main-add-image-button'
-            type='button'
+            className="ap-form-main-add-image-button"
+            type="button"
             onClick={handleImageInputClick}
             disabled={loading}
           >
-            <p className='ap-form-main-add-image-button-text'>
+            <p className="ap-form-main-add-image-button-text">
               Thêm vào bài viết của bạn
             </p>
-            <p className='ap-form-main-add-image-button-icons'>
-              <FaImage className='ap-form-main-add-image-button-icons-image' />
-              <FaMapMarkerAlt className='ap-form-main-add-image-button-icons-map' />
+            <p className="ap-form-main-add-image-button-icons">
+              <FaImage className="ap-form-main-add-image-button-icons-image" />
+              <FaMapMarkerAlt className="ap-form-main-add-image-button-icons-map" />
             </p>
           </button>
           <input
-            type='file'
-            accept='image/*'
+            type="file"
+            accept="image/*"
             ref={imageInputRef}
             hidden
             onChange={handleImageChange}
           />
           <Button
-            type='primary'
-            className='ap-form-main-submit'
+            type="primary"
+            className="ap-form-main-submit"
             onClick={handleSubmit}
             disabled={disabled || loading}
           >
