+++ frontend/src/views/pages/individual/SmallPost/index.jsx
@@ -8,11 +8,10 @@ import {
   HeartFilled,
   HeartOutlined,
   PushpinFilled,
-  PushpinOutlined,
 } from "@ant-design/icons";
 import { FaCircle, FaRegCommentDots } from "react-icons/fa";
 import TextArea from "antd/es/input/TextArea";
-import { ImGrin, ImSpinner } from "react-icons/im";
+import { ImGrin } from "react-icons/im";
 import { BsCursorFill } from "react-icons/bs";
 import SmallComment from "../SmallComment";
 import { BiDotsHorizontal } from "react-icons/bi";
@@ -28,8 +27,6 @@ function SmallPost(props) {
   const [isShowComment, setIsShowComment] = useState(false);
 
   const [comments, setComments] = useState(post.comments);
-
-  const [isPin, setIsPin] = useState(false);
 
   const handleNavigateEdit = () => {
     navigate("/editpost", { state: { post, user } }); // Đường dẫn đến trang bạn muốn chuyển hướng
@@ -49,13 +46,7 @@ function SmallPost(props) {
   const onChangeComment = (e) => {
     setComment(e.target.value);
   };
-
-  const onPin = () => {
-    setIsPin(!isPin);
-  };
-
   const navigate = useNavigate();
-
   const items = [
     {
       label: (
@@ -102,22 +93,22 @@ function SmallPost(props) {
   ];
 
   return (
-    <div className="group-small-post-container">
-      <div className="group-small-post-user-info-container">
+    <div className="individual-small-post-container">
+      <div className="individual-small-post-user-info-container">
         <div className="d-flex">
           <img
-            className="group-small-post-avatar"
+            className="individual-small-post-avatar"
             alt="user avatar"
             src={user.avatar}
           />
-          <div className="group-small-post-info-container">
+          <div className="individual-small-post-info-container">
             <div className="d-flex align-items-center">
-              <p class="group-small-post-username">{user.displayName}</p>
-              <p class="group-small-post-status">
+              <p class="individual-small-post-username">{user.displayName}</p>
+              <p class="individual-small-post-status">
                 <FaCircle />
               </p>
             </div>
-            <p className="group-small-post-createdAt">{post.createdAt}</p>
+            <p className="individual-small-post-createdAt">{post.createdAt}</p>
           </div>
         </div>
         <div>
@@ -126,22 +117,22 @@ function SmallPost(props) {
               items,
             }}
             trigger={["click"]}
-            className="group-small-post-dropdown"
+            className="individual-small-post-dropdown"
           >
             <BiDotsHorizontal onClick={(e) => e.preventDefault()} />
           </Dropdown>
         </div>
       </div>
-      <div className="group-small-post-content-container">
-        <p class="group-small-post-content">{post.content}</p>
+      <div className="individual-small-post-content-container">
+        <p class="individual-small-post-content">{post.content}</p>
         {post.image && (
           <img
-            className="group-small-post-image"
+            className="individual-small-post-image"
             alt="post pic"
             src={post.image}
           />
         )}
-        <div className="group-small-post-like-comment-container">
+        <div className="individual-small-post-like-comment-container">
           <p>{post.likeNum}</p>
           <Button
             icon={
@@ -152,57 +143,49 @@ function SmallPost(props) {
               )
             }
             onClick={onLike}
-            className="group-small-post-button"
+            className="individual-small-post-button"
             style={{ color: "#EB3223" }}
           />
           <p>{post.commentNum}</p>
           <Button
             icon={<FaRegCommentDots />}
             onClick={onClickComment}
-            className="group-small-post-button"
-          />
-          <Button
-            icon={
-              isPin ? (
-                <PushpinFilled style={{ fontSize: "25px", color: "#ebc351" }} />
-              ) : (
-                <PushpinOutlined style={{ fontSize: "25px" }} />
-              )
-            }
-            className="group-small-post-button"
-            onClick={onPin}
+            className="individual-small-post-button"
           />
         </div>
         {isShowComment && (
-          <div className="group-small-post-comment-container">
+          <div className="individual-small-post-comment-container">
             <img
               alt="user avatar"
               src={user.avatar}
-              className="group-small-post-avatar"
+              className="individual-small-post-avatar"
             />
             <TextArea
               placeholder={`Bình luận với vai trò ${user.displayName}`}
-              className="group-small-post-input"
+              className="individual-small-post-input"
               autoSize
               value={comment}
               onChange={(value) => onChangeComment(value)}
             />
-            <div className="group-small-post-button-container">
+            <div className="individual-small-post-button-container">
               <Button
                 icon={<FaRegImage />}
-                className="group-small-post-button"
+                className="individual-small-post-button"
               />
-              <Button icon={<ImGrin />} className="group-small-post-button" />
+              <Button
+                icon={<ImGrin />}
+                className="individual-small-post-button"
+              />
               <Button
                 icon={<BsCursorFill />}
-                className="group-small-post-button"
+                className="individual-small-post-button"
               />
             </div>
           </div>
         )}
       </div>
       {isShowComment && (
-        <div className="group-small-post-comment-wrapper">
+        <div className="individual-small-post-comment-wrapper">
           {comments &&
             comments.length > 0 &&
             comments.map((comment) => (
