+++ frontend/src/views/pages/individual/SmallPost/index.scss
@@ -1,6 +1,6 @@
 @import "../../../../constants/color.scss";
 
-.group-small-post {
+.individual-small-post {
   &-container {
     margin: 10px 15px;
     background-color: $light_gray;
@@ -17,7 +17,7 @@
     border-radius: 50%;
     object-fit: cover;
   }
-  &-dropdown {
+  &-dropdown{
     cursor: pointer;
     font-size: 25px;
   }
@@ -29,16 +29,16 @@
     margin-left: 10px;
     align-items: center;
   }
-  &-createdAt {
-    color: $subText;
+  &-createdAt{
+    color : $subText;
   }
-  &-username {
+  &-username{
     font-size: 22px;
     font-weight: bold;
     margin-right: 10px !important;
   }
-  &-status {
-    color: $light-green;
+  &-status{
+    color : $light-green;
   }
   &-button {
     background-color: transparent !important;
@@ -60,25 +60,24 @@
     height: 40px !important;
     padding-top: 5px;
   }
-  &-like-comment-container {
+  &-like-comment-container{
     display: flex;
     align-items: center;
-    margin: 10px 0px;
-    p {
+    margin : 10px 0px;
+    p{
       margin: 0;
       padding: 0;
-      margin: 0px 5px;
+      margin : 0px 5px
     }
   }
   &-content {
-    padding: 0;
-    margin: 10px 0px;
+    padding : 0;
+    margin : 10px 0px;
   }
   &-image {
     height: 400px;
   }
-  &-content-container,
-  &-comment-wrapper {
+  &-content-container, &-comment-wrapper{
     padding-left: 30px;
   }
   &-comment-container {
